// angularapp/src/app/model/employeeMonthlyAdvance.ts
export interface EmployeeMonthlyAdvance {
    ID:number;
    EMP_ID: number;
    EMP_NAME: string | null;
    EMP_ROLE:string | null;
    EMP_IC_NEW: string | null;
    EMP_IC_OLD: string | null;
    EMP_PASSPORT_NO: string | null;
    EMPFL_BANK: string | null;
    EMPFL_BK_ACCNO: string | null;
    PAYMODE: string | null;
}